MdeModulePkg/MemoryTypeInfoSecVarCheckLib: Elaborate on expected mem types.
This variable check library instance currently checks the size of the
`MemoryTypeInformation` UEFI variable to determine if it contains the
total size expected based on the expected number of entries and the
`EFI_MEMORY_TYPE_INFORMATION` structure size where the structure is
used to represent an individual entry.

This changes adds a comment near entry count to describe what entries
are expected and updates the associated assert message to indicate the
developer should check the actual memory types against the expected
memory types if a size mismatch occurs.