MdeModulePkg/UefiBootManagerLib: Register var policy on notify (#1276)
## Description

Variable policy is currently set in the constructor. In some system
designs, variable policy might not be running when the constructor of
this library is executed. To make the policy registration process more
robust, move registration to a protocol notify.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?
- [x] Backport to release branch?

## How This Was Tested

- Verify notify is called on variable policy protocol installation and
register policy successfully.
- Verify subsequent calls to the protocol gracefully exit by recognizing
`EFI_ALREADY_STARTED` is returned from `RegisterBasicVariablePolicy()`.

## Integration Instructions

N/A

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>